sync: move SyncProgressEstimator to dcrlibwallet package
- fix error that occurs when the shutdown process tries to close and already closed channel
- Remove sync cancellation from shutdown process since the wallet loader already handles that
