+++ go.mod
@@ -1,9 +1,16 @@
 module github.com/raedahgroup/dcrlibwallet
 
 require (
+	4d63.com/gochecknoglobals v0.0.0-20190306162314-7c3491d2b6ec // indirect
+	4d63.com/gochecknoinits v0.0.0-20180528051558-14d5915061e5 // indirect
 	github.com/AndreasBriese/bbloom v0.0.0-20190306092124-e2d15f34fcf9 // indirect
+	github.com/BurntSushi/toml v0.3.1 // indirect
 	github.com/DataDog/zstd v1.3.5 // indirect
+	github.com/alecthomas/gocyclo v0.0.0-20150208221726-aa8f8b160214 // indirect
+	github.com/alexflint/go-arg v1.0.0 // indirect
+	github.com/alexkohler/nakedret v0.0.0-20190321114339-98ae56e4e0f3 // indirect
 	github.com/asdine/storm v0.0.0-20190216191021-fe89819f6282
+	github.com/client9/misspell v0.3.4 // indirect
 	github.com/decred/dcrd/addrmgr v1.0.2
 	github.com/decred/dcrd/blockchain/stake v1.1.0
 	github.com/decred/dcrd/chaincfg v1.3.0
@@ -28,8 +35,25 @@ require (
 	github.com/decred/slog v1.0.0
 	github.com/dgraph-io/badger v1.5.4
 	github.com/dgryski/go-farm v0.0.0-20190104051053-3adb47b1fb0f // indirect
+	github.com/gordonklaus/ineffassign v0.0.0-20180909121442-1003c8bd00dc // indirect
+	github.com/jgautheron/goconst v0.0.0-20170703170152-9740945f5dcb // indirect
 	github.com/jrick/logrotate v1.0.0
+	github.com/kisielk/errcheck v1.2.0 // indirect
+	github.com/kisielk/gotool v1.0.0 // indirect
+	github.com/mdempsky/maligned v0.0.0-20180708014732-6e39bd26a8c8 // indirect
+	github.com/mdempsky/unconvert v0.0.0-20190325185700-2f5dc3378ed3 // indirect
+	github.com/mibk/dupl v1.0.0 // indirect
+	github.com/opennota/check v0.0.0-20180911053232-0c771f5545ff // indirect
 	github.com/pkg/errors v0.8.1 // indirect
+	github.com/securego/gosec v0.0.0-20190517133546-0ebfa2f8b7f8 // indirect
+	github.com/stripe/safesql v0.0.0-20171221195208-cddf355596fe // indirect
+	github.com/tsenart/deadcode v0.0.0-20160724212837-210d2dc333e9 // indirect
+	github.com/walle/lll v0.0.0-20160702150637-8b13b3fbf731 // indirect
 	go.etcd.io/bbolt v1.3.0
+	golang.org/x/tools v0.0.0-20190521171243-7927dbab1be7 // indirect
 	google.golang.org/appengine v1.5.0 // indirect
+	honnef.co/go/tools v0.0.0-20190418001031-e561f6794a2a // indirect
+	mvdan.cc/interfacer v0.0.0-20180901003855-c20040233aed // indirect
+	mvdan.cc/lint v0.0.0-20170908181259-adc824a0674b // indirect
+	mvdan.cc/unparam v0.0.0-20190310220240-1b9ccfa71afe // indirect
 )
